seperate some scraper code to util file
